requirements.txt
fastapi==0.115.14
gradio
requests==2.32.4
# Remove llama-cpp-python and langchain, not needed for cloud LLM
# llama-cpp-python
# langchain
# langchain_core
sentencepiece==0.2.0
mflux==0.9.0
uvicorn==0.30.0
pydantic==2.7.1
# Add HuggingFace Transformers and Diffusers for LLM and image models
transformers==4.41.2
diffusers==0.27.2
torch==2.3.0
# For image processing
Pillow==10.3.0